+++ code/Test/test2.c
@@ -1,16 +1,3 @@
-#include <stdio.h>
-
-void printS(char string[])
-{
-    int c = 0;
-    while (string[c] != '\0')
-    {
-        printf("%c", string[c]);
-        c++;
-    }
-    printf("\n");
-}
-
 int lenS(char string[])
 {
     int c = 0;
@@ -18,7 +5,8 @@ int lenS(char string[])
     return c;
 }
 
-char changename()
+
+int main(int argc, char const *argv[])
 {
     char PlayerX[16] = "NOME DO JOGADOR X";
     int len = lenS(PlayerX);
@@ -29,12 +17,6 @@ char changename()
     {
         *(PlayerX+i) = name[i];
     }
-    //printf("%s\n", PlayerX);
-    return PlayerX;
-}
-int main()
-{
-    char nome[16];
-    *nome = changename();
-    printS(nome);
+    printf("%s\n", PlayerX);
+    return 0;
 }